requirements.txt
librosa==0.10.2.post1
numpy==1.24.4
Pillow==11.1.0
playsound==1.3.0
pytest==8.3.4
sounddevice==0.5.1
soundfile==0.13.0
torch==2.4.1+cu118
webrtcvad==2.0.10
